requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.1
django-colorfield==0.7.1
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
django-smart-selects==1.5.9
djangorestframework==3.13.1
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
imagecolor==1.2.1
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
Pillow==9.1.1
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.8.0
